# 04-API_SPEC.yaml
openapi: 3.1.0
...
info:
  title: ProofPay Application API
  version: 0.1.0
  description: 'ProofPay compiles the review queue between ''work submitted'' and ''payment released''
    into executable acceptance checks — and pays only when evidence passes.


    Phase 2 contract, not an implemented server. Read docs/00-PROJECT_CONTEXT.md and docs/02-PRD.md. One
    agency, one fixture, two sandbox recipients, USD, exactly three approved checks and three families.
    Examples/identifiers are illustrative unless execution evidence says otherwise. Cookie domain mutations
    require CSRF and Idempotency-Key; internal tool/runner credentials never enter model context. Paid
    requires a matched actual PayPal item SUCCESS.'
  license:
    name: MIT
    identifier: MIT
servers:
- url: http://localhost:8000
  description: Local Compose API; production cookies require HTTPS.
...
security:
- SessionCookie: []
paths:
...
  /api/tasks:
    get:
      operationId: list_tasks
      summary: list tasks
      description: AG Grid work queue. Defaults to current demo run plus its labeled genuine completed
        archive. An explicit demo_run_id filter lists only that run. Archive rows retain original task,
        obligation and provider IDs; other historical records remain inspectable.
      tags:
      - Tasks
      x-roles:
      - owner
      x-requirement-ids:
      - FR-15
      - FR-35
      responses:
        '200':
          description: Success
          headers:
            X-Request-ID:
              schema: *id001
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskList'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/Invalid'
        '429':
          $ref: '#/components/responses/RateLimited'
        '503':
          $ref: '#/components/responses/Unavailable'
        default:
          $ref: '#/components/responses/Unexpected'
      security:
      - SessionCookie: []
      parameters:
      - $ref: '#/components/parameters/Cursor'
      - $ref: '#/components/parameters/Limit'
      - name: state
        in: query
        schema:
          type: string
          enum:
          - brief_captured
          - checks_approved
          - awaiting_delivery
          - verifying
          - correction_requested
          - evidence_passed
          - payment_initiated
          - reconciling
          - paid
          - failed
          - unclaimed
          - cancelled
      - name: family
        in: query
        schema: &id003
          type: string
          enum:
          - responsive_css
          - api_endpoint
          - keyboard_accessibility
      - name: recipient_ref
        in: query
        schema:
          type: string
          enum:
          - contractor_maya
          - contractor_leo
      - name: sort
        in: query
        schema:
          type: string
          enum:
          - updated_desc
          - created_desc
          - amount_asc
      - name: demo_run_id
        in: query
        schema: *id001
  /api/tasks/{task_id}:
    get:
      operationId: get_task
      summary: get task
      description: ''
      tags:
      - Tasks
      x-roles:
      - owner
      - contractor
      x-requirement-ids:
      - FR-15
      - FR-34
      responses:
        '200':
          description: Success
          headers:
            X-Request-ID:
              schema: *id001
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/Invalid'
        '429':
          $ref: '#/components/responses/RateLimited'
        '503':
          $ref: '#/components/responses/Unavailable'
        default:
          $ref: '#/components/responses/Unexpected'
      security:
      - SessionCookie: []
      parameters:
      - name: task_id
        in: path
        required: true
        schema: *id001
...
  /api/tasks/{task_id}/receipt:
    get:
      operationId: get_receipt
      summary: get receipt
      description: ''
      tags:
      - Receipts
      x-roles:
      - owner
      - contractor
      x-requirement-ids:
      - FR-13
      - FR-33
      responses:
        '200':
          description: Success
          headers:
            X-Request-ID:
              schema: *id001
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Receipt'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/Invalid'
        '429':
          $ref: '#/components/responses/RateLimited'
        '503':
          $ref: '#/components/responses/Unavailable'
        default:
          $ref: '#/components/responses/Unexpected'
      security:
      - SessionCookie: []
      parameters:
      - name: task_id
        in: path
        required: true
        schema: *id001
...
  /api/contractor/tasks:
    get:
      operationId: list_contractor_tasks
      summary: list contractor tasks
      description: Fixed to session recipient; no recipient override parameter.
      tags:
      - Contractor
      x-roles:
      - contractor
      x-requirement-ids:
      - FR-15
      - FR-34
      responses:
        '200':
          description: Success
          headers:
            X-Request-ID:
              schema: *id001
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskList'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/Invalid'
        '429':
          $ref: '#/components/responses/RateLimited'
        '503':
          $ref: '#/components/responses/Unavailable'
        default:
          $ref: '#/components/responses/Unexpected'
      security:
      - SessionCookie: []
      parameters:
      - $ref: '#/components/parameters/Cursor'
      - $ref: '#/components/parameters/Limit'
  /api/contractor/tasks/{task_id}:
    get:
      operationId: get_contractor_task
      summary: get contractor task
      description: ''
      tags:
      - Contractor
      x-roles:
      - contractor
      x-requirement-ids:
      - FR-34
      responses:
        '200':
          description: Success
          headers:
            X-Request-ID:
              schema: *id001
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/Invalid'
        '429':
          $ref: '#/components/responses/RateLimited'
        '503':
          $ref: '#/components/responses/Unavailable'
        default:
          $ref: '#/components/responses/Unexpected'
      security:
      - SessionCookie: []
      parameters:
      - name: task_id
        in: path
        required: true
        schema: *id001
...
components:
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: proofpay_session
      description: Opaque session; role/agency resolved server-side.
    WorkerServiceToken:
      type: http
      scheme: bearer
      description: Application worker audience; job/stage/lease context checked. Never a PayPal token.
    RunnerServiceToken:
      type: http
      scheme: bearer
      description: Distinct runner audience; no app financial or AI access.
  parameters:
...
    Cursor:
      name: cursor
      in: query
      schema:
        type: string
        maxLength: 256
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 50
...
  responses:
    BadRequest:
      description: Malformed JSON/header or duplicate keys.
      headers:
        X-Request-ID:
          schema: *id001
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Missing/expired session or service credential.
      headers:
        X-Request-ID:
          schema: *id001
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Role, CSRF, service audience or stage denied.
      headers:
        X-Request-ID:
          schema: *id001
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource absent or inaccessible in agency/contractor scope.
      headers:
        X-Request-ID:
          schema: *id001
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    Invalid:
      description: Schema, scope, parameter, money or authority input invalid.
      headers:
        X-Request-ID:
          schema: *id001
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    RateLimited:
      description: Request limit reached; retry after the supplied delay.
      headers:
        X-Request-ID:
          schema: *id001
        Retry-After:
          schema:
            type: string
            maxLength: 80
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unavailable:
      description: Durable storage or required service unavailable.
      headers:
        X-Request-ID:
          schema: *id001
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unexpected:
      description: Unexpected error; use request_id without exposed secrets.
      headers:
        X-Request-ID:
          schema: *id001
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Money:
      type: object
      additionalProperties: false
      properties:
        currency:
          type: string
          enum:
          - USD
        value:
          type: string
          maxLength: 16
          pattern: ^(0|[1-9][0-9]{0,8})\.[0-9]{2}$
      required:
      - currency
      - value
    PositiveMoney:
      type: object
      additionalProperties: false
      properties:
        currency:
          type: string
          enum:
          - USD
        value:
          type: string
          maxLength: 16
          pattern: ^(?!0\.00$)(0|[1-9][0-9]{0,8})\.[0-9]{2}$
      required:
      - currency
      - value
    PaymentTerms:
      type: object
      additionalProperties: false
      properties:
        recipient_ref:
          type: string
          enum:
          - contractor_maya
          - contractor_leo
        amount:
          $ref: '#/components/schemas/PositiveMoney'
        principal_cap:
          $ref: '#/components/schemas/PositiveMoney'
        max_attempts:
          type: integer
          minimum: 1
          maximum: 3
        expires_at: &id002
          type: string
          format: date-time
          pattern: .*(Z|[+-][0-9]{2}:[0-9]{2})$
      required:
      - recipient_ref
      - amount
      - principal_cap
      - max_attempts
      - expires_at
      description: Cap must cover amount; expiry must be future at approval. Workspace allowance remains
        an independent durable control.
    Check_viewport_no_horizontal_overflow:
      type: object
      additionalProperties: false
      properties:
        check_id:
          type: string
          enum:
          - C01
          - C02
          - C03
        type:
          type: string
          const: viewport_no_horizontal_overflow
        params:
          type: object
          additionalProperties: false
          properties:
            width:
              type: integer
              enum:
              - 320
            target_ref:
              type: string
              enum:
              - checkout
          required:
          - width
          - target_ref
        compiled_by:
          type: string
          enum:
          - ai
        approved:
          type: boolean
      required:
      - check_id
      - type
      - params
      - compiled_by
      - approved
...
    AcceptanceCheck:
      oneOf:
      - $ref: '#/components/schemas/Check_viewport_no_horizontal_overflow'
      - $ref: '#/components/schemas/Check_cart_total_unchanged'
      - $ref: '#/components/schemas/Check_keyboard_checkout_reachable'
      - $ref: '#/components/schemas/Check_api_status'
      - $ref: '#/components/schemas/Check_api_schema'
      - $ref: '#/components/schemas/Check_api_total_matches_fixture'
      - $ref: '#/components/schemas/Check_keyboard_activation'
      - $ref: '#/components/schemas/Check_accessible_control_name'
...
    Error:
      type: object
      additionalProperties: false
      properties:
        code:
          type: string
          maxLength: 80
        message:
          type: string
          maxLength: 1000
        request_id: *id001
        details:
          type: array
          items:
            type: object
            additionalProperties: false
            properties:
              field:
                oneOf:
                - type: string
                  maxLength: 200
                - type: 'null'
              reason:
                type: string
                maxLength: 600
            required:
            - field
            - reason
          minItems: 0
          maxItems: 30
        retryable:
          type: boolean
      required:
      - code
      - message
      - request_id
      - details
      - retryable
      description: Codes include INVALID_REQUEST, UNSUPPORTED_SCOPE, UNAUTHORIZED, FORBIDDEN, NOT_FOUND,
        STALE_REVISION, IDEMPOTENCY_CONFLICT, WORKSPACE_BUSY, AUTHORITY_EXPIRED, GUARD_FAILED and DEPENDENCY_UNAVAILABLE.
        No secrets or raw receiver values.
...
    Contractor:
      type: object
      additionalProperties: false
      properties:
        id: *id001
        recipient_ref:
          type: string
          enum:
          - contractor_maya
          - contractor_leo
        display_name:
          type: string
          maxLength: 120
        environment:
          type: string
          enum:
          - sandbox
      required:
      - id
      - recipient_ref
      - display_name
      - environment
...
    Brief:
      type: object
      additionalProperties: false
      properties:
        id: *id001
        task_id: *id001
        revision_id: *id001
        revision:
          type: integer
          minimum: 1
        digest: *id004
        title:
          type: string
          maxLength: 160
        text:
          type: string
          maxLength: 4000
        family: *id003
        fixture_ref:
          type: string
          enum:
          - checkout_fixture
        terms:
          $ref: '#/components/schemas/PaymentTerms'
        created_at: *id002
        updated_at: *id002
      required:
      - id
      - task_id
      - revision_id
      - revision
      - digest
      - title
      - text
      - family
      - fixture_ref
      - terms
      - created_at
      - updated_at
...
    MandatePayload:
      type: object
      additionalProperties: false
      properties:
        mandate_id: *id001
        version:
          type: integer
          minimum: 1
        status:
          type: string
          enum:
          - draft
          - approved
        payer:
          type: object
          additionalProperties: false
          properties:
            account_ref:
              type: string
              enum:
              - payer_sandbox_us
          required:
          - account_ref
        payee:
          type: object
          additionalProperties: false
          properties:
            recipient_ref:
              type: string
              enum:
              - contractor_maya
              - contractor_leo
            display_name:
              type: string
              maxLength: 120
            email_hash: *id004
          required:
          - recipient_ref
          - display_name
          - email_hash
        amount:
          $ref: '#/components/schemas/PositiveMoney'
        budget:
          type: object
          additionalProperties: false
          properties:
            max_total:
              type: string
              maxLength: 16
              pattern: ^(0|[1-9][0-9]{0,8})\.[0-9]{2}$
            max_attempts:
              type: integer
              minimum: 1
              maximum: 3
          required:
          - max_total
          - max_attempts
        authority:
          type: object
          additionalProperties: false
          properties:
            type:
              type: string
              enum:
              - conditional_release
            condition:
              type: string
              enum:
              - all checks pass
            expires_at: *id002
          required:
          - type
          - condition
          - expires_at
        acceptance_checks:
          type: array
          items:
            $ref: '#/components/schemas/AcceptanceCheck'
          minItems: 3
          maxItems: 3
        audit:
          type: object
          additionalProperties: false
          properties:
            approved_by:
              oneOf:
              - *id001
              - type: 'null'
            approved_at:
              oneOf:
              - *id002
              - type: 'null'
            immutable:
              type: boolean
          required:
          - approved_by
          - approved_at
          - immutable
      required:
      - mandate_id
      - version
      - status
      - payer
      - payee
      - amount
      - budget
      - authority
      - acceptance_checks
      - audit
      description: Seeded mandate skeleton. Approved snapshot status/audit remain frozen; current lifecycle
        state is outside this payload. Protected receiver is never public.
    MandateVersion:
      type: object
      additionalProperties: false
      properties:
        id: *id001
        mandate_id: *id001
        version:
          type: integer
          minimum: 1
        lifecycle_state:
          type: string
          enum:
          - draft
          - approved
          - expired
          - superseded
          - exhausted
        digest: *id004
        snapshot:
          $ref: '#/components/schemas/MandatePayload'
        created_at: *id002
      required:
      - id
      - mandate_id
      - version
      - lifecycle_state
      - digest
      - snapshot
      - created_at
...
    Task:
      type: object
      additionalProperties: false
      properties:
        id: *id001
        agency_id: *id001
        demo_run_id: *id001
        is_featured_archive:
          type: boolean
        brief_id: *id001
        title:
          type: string
          maxLength: 160
        family: *id003
        contractor:
          $ref: '#/components/schemas/Contractor'
        amount:
          $ref: '#/components/schemas/PositiveMoney'
        state:
          type: string
          enum:
          - brief_captured
          - checks_approved
          - awaiting_delivery
          - verifying
          - correction_requested
          - evidence_passed
          - payment_initiated
          - reconciling
          - paid
          - failed
          - unclaimed
          - cancelled
        hold_reasons:
          type: array
          items:
            type: string
            maxLength: 160
          minItems: 0
          maxItems: 12
        review_required:
          type: boolean
        version:
          type: integer
          minimum: 1
        mandate_version_id:
          oneOf:
          - *id001
          - type: 'null'
        current_delivery_id:
          oneOf:
          - *id001
          - type: 'null'
        current_bundle_id:
          oneOf:
          - *id001
          - type: 'null'
        existing_attempt_id:
          oneOf:
          - *id001
          - type: 'null'
        allowed_actions:
          type: array
          items:
            type: string
            enum:
            - compile
            - approve
            - submit_delivery
            - resolve_review
            - request_payout
            - reconcile_payment
            - retry_payment
            - cancel_unclaimed
            - reset
            - replay
          minItems: 0
          maxItems: 10
        updated_at: *id002
      required:
      - id
      - agency_id
      - demo_run_id
      - is_featured_archive
      - brief_id
      - title
      - family
      - contractor
      - amount
      - state
      - hold_reasons
      - review_required
      - version
      - mandate_version_id
      - current_delivery_id
      - current_bundle_id
      - existing_attempt_id
      - allowed_actions
      - updated_at
...
    Delivery:
      type: object
      additionalProperties: false
      properties:
        id: *id001
        task_id: *id001
        mandate_version_id: *id001
        artifact_version_id: *id001
        artifact_digest: *id004
        claim:
          type: string
          maxLength: 4000
        claim_ref: &id005
          type: string
          maxLength: 64
          pattern: ^(claim|result|screenshot|manifest):[0-9a-f]{8}(-[0-9a-f]{4}){3}-[0-9a-f]{12}$
        sequence:
          type: integer
          minimum: 1
        previous_delivery_id:
          oneOf:
          - *id001
          - type: 'null'
        is_current:
          type: boolean
        verification_job_id: *id001
        created_at: *id002
      required:
      - id
      - task_id
      - mandate_version_id
      - artifact_version_id
      - artifact_digest
      - claim
      - claim_ref
      - sequence
      - previous_delivery_id
      - is_current
      - verification_job_id
      - created_at
...
    GuardDecision:
      type: object
      additionalProperties: false
      properties:
        code:
          type: string
          maxLength: 80
        passed:
          type: boolean
        reason:
          oneOf:
          - type: string
            maxLength: 600
          - type: 'null'
      required:
      - code
      - passed
      - reason
    BudgetView:
      type: object
      additionalProperties: false
      properties:
        workspace_limit:
          $ref: '#/components/schemas/Money'
        workspace_reserved:
          $ref: '#/components/schemas/Money'
        workspace_consumed:
          $ref: '#/components/schemas/Money'
        workspace_available:
          $ref: '#/components/schemas/Money'
        mandate_principal_cap:
          $ref: '#/components/schemas/Money'
        reserved_principal:
          $ref: '#/components/schemas/Money'
        consumed_principal:
          $ref: '#/components/schemas/Money'
        released_principal:
          $ref: '#/components/schemas/Money'
        remaining_business_attempts:
          type: integer
          minimum: 0
          maximum: 3
      required:
      - workspace_limit
      - workspace_reserved
      - workspace_consumed
      - workspace_available
      - mandate_principal_cap
      - reserved_principal
      - consumed_principal
      - released_principal
      - remaining_business_attempts
      description: Configured principal accounting, not provider wallet balance. Fees are separate.
    PaymentAttempt:
      type: object
      additionalProperties: false
      properties:
        id: *id001
        obligation_id: *id001
        task_id: *id001
        mandate_version_id: *id001
        attempt_no:
          type: integer
          minimum: 1
          maximum: 3
        state:
          type: string
          enum:
          - prepared
          - dispatching
          - unknown
          - processing
          - unclaimed
          - success
          - failed
          - cancelled
          - blocked
        item_state:
          type: string
          enum:
          - created
          - processing
          - success
          - unclaimed
          - cancelled
          - failed
          - blocked
        principal:
          $ref: '#/components/schemas/PositiveMoney'
        observed_fee:
          oneOf:
          - $ref: '#/components/schemas/Money'
          - type: 'null'
        sender_batch_id:
          type: string
          maxLength: 256
        sender_item_id:
          type: string
          maxLength: 63
        batch_id:
          oneOf:
          - type: string
            maxLength: 128
          - type: 'null'
        item_id:
          oneOf:
          - type: string
            maxLength: 128
          - type: 'null'
        transaction_id:
          oneOf:
          - type: string
            maxLength: 128
          - type: 'null'
        raw_status:
          oneOf:
          - type: string
            maxLength: 80
          - type: 'null'
        unresolved:
          type: boolean
        nonpayment_confirmed:
          type: boolean
        guard_decisions:
          type: array
          items:
            $ref: '#/components/schemas/GuardDecision'
          minItems: 0
          maxItems: 16
        hold_reasons:
          type: array
          items:
            type: string
            maxLength: 160
          minItems: 0
          maxItems: 12
        allowed_actions:
          type: array
          items:
            type: string
            enum:
            - reconcile_payment
            - retry_payment
            - cancel_unclaimed
          minItems: 0
          maxItems: 3
        last_reconciled_at:
          oneOf:
          - *id002
          - type: 'null'
        created_at: *id002
      required:
      - id
      - obligation_id
      - task_id
      - mandate_version_id
      - attempt_no
      - state
      - item_state
      - principal
      - observed_fee
      - sender_batch_id
      - sender_item_id
      - batch_id
      - item_id
      - transaction_id
      - raw_status
      - unresolved
      - nonpayment_confirmed
      - guard_decisions
      - hold_reasons
      - allowed_actions
      - last_reconciled_at
      - created_at
...
    ReceiptEvent:
      type: object
      additionalProperties: false
      properties:
        id: *id001
        type:
          type: string
          maxLength: 80
        at: *id002
        actor_label:
          type: string
          maxLength: 150
        summary:
          type: string
          maxLength: 1000
        source_id: *id001
        evidence_refs:
          type: array
          items: *id005
          minItems: 0
          maxItems: 20
      required:
      - id
      - type
      - at
      - actor_label
      - summary
      - source_id
      - evidence_refs
    Receipt:
      type: object
      additionalProperties: false
      properties:
        task_id: *id001
        state:
          type: string
          enum:
          - brief_captured
          - checks_approved
          - awaiting_delivery
          - verifying
          - correction_requested
          - evidence_passed
          - payment_initiated
          - reconciling
          - paid
          - failed
          - unclaimed
          - cancelled
        label:
          type: string
          enum:
          - payment_record
          - paid_receipt
          - returned_record
        brief:
          $ref: '#/components/schemas/Brief'
        mandate_version:
          $ref: '#/components/schemas/MandateVersion'
        delivery:
          oneOf:
          - $ref: '#/components/schemas/Delivery'
          - type: 'null'
        evidence_bundle_id:
          oneOf:
          - *id001
          - type: 'null'
        payment:
          oneOf:
          - $ref: '#/components/schemas/PaymentAttempt'
          - type: 'null'
        budget:
          $ref: '#/components/schemas/BudgetView'
        timeline:
          type: array
          items:
            $ref: '#/components/schemas/ReceiptEvent'
          minItems: 0
          maxItems: 100
        last_reconciled_at:
          oneOf:
          - *id002
          - type: 'null'
      required:
      - task_id
      - state
      - label
      - brief
      - mandate_version
      - delivery
      - evidence_bundle_id
      - payment
      - budget
      - timeline
      - last_reconciled_at
...
    TaskList:
      type: object
      additionalProperties: false
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Task'
          minItems: 0
          maxItems: 100
        next_cursor:
          oneOf:
          - type: string
            maxLength: 256
          - type: 'null'
      required:
      - items
      - next_cursor
...